a lot of changes...
semi-functional; still missing the entire network aspect which is its purposes, however ioctl's have been implemented as has been mmap et al. I still need to add poll() which is pretty trivial. Then the network hooks. IPv6 support is totally broken at present and won't actually be added until ipv4 is working; its going to require a bit more thought because even represented as a bit i cant realistically fit large chunks of the ipv6 address space into memory and so it will have to be segmented down further in most instances-- it is the primary rational for using poll() which is expected to take longer.

more ioctl's will probably be added; for instance after a scan completes it will probably be desirable to say "rescan the same network but only transmit to hosts we didn't hear back from".